rename : rename StudentAddDailog, GradeChips,

+++ app/src/main/java/com/example/stmt/data/student/Student.kt
@@ -20,7 +20,7 @@ data class Student(
 
     var tel: String,
 
-    var classId: Long
+    var classId: Int
 
 
 ) {
